+++ src/Contracts/DtoContract.php
@@ -10,6 +10,10 @@ interface DtoContract
 
     public function except(string ...$keys): DtoContract;
 
+    public function with(string $key, $value): DtoContract;
+
+    public function override(string $key, $value): DtoContract;
+
     public function toArray(): array;
 
     public function isImmutable(): bool;