+++ CHANGELOG.md
@@ -2,13 +2,17 @@
 
 All notable changes to `data-transfer-object` will be documented in this file
 
+## 2.1.0" - 2019-04-18
+
+- Added support to override or add attributes on non immutable dto's
+
 ## 2.0.1" - 2019-04-15
 
 - Improved error messages
 
 ## 2.0.0" - 2019-04-15
 
-- Optional properties support + major immutability improvements
+- new package release (namespace changed). Optional properties support + major immutability improvements
 
 ## 1.8.0 - 2019-03-18